pyproject.toml
[tool.poetry]
name = "hive-daily-power-up"
version = "1.3.0"
description = "Daily power up a Hive account"
authors = ["Brian of London (home imac) <brian@v4v.app>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
single-source = "^0.3.0"
beem = {git = "https://github.com/brianoflondon/beem.git"}
python-dotenv = "^1.0.0"
colorlog = "^6.8.0"
termcolor = "^2.4.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
